feat: 시스템 롤 삭제 개발

+++ src/main/java/com/cjw/reactivecommunityproject/web/system/role_management/model/entity/SystemRoleManagementInsertEntity.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record SystemRoleManagementInsertEntity(
-        Long uid,
+        Integer uid,
         String name,
         String userUid
 ) {
