+++ src/main/java/com/cjw/reactivecommunityproject/web/system/role_management/model/request/SystemRoleManagementCreateVO.java
@@ -8,7 +8,7 @@ import lombok.Builder;
 public record SystemRoleManagementCreateVO(
         @Min(value = 1, groups = {SystemRoleManagementValidationGroup.Create.class}, message = "uid 은 1 이상이어야 합니다.")
         @Max(value = 127, groups = {SystemRoleManagementValidationGroup.Create.class}, message = "uid 은 128 미만이어야 합니다.")
-        Long uid,
+        Integer uid,
         @NotBlank(groups = {SystemRoleManagementValidationGroup.Create.class}, message = "name 은 공백일 수 없습니다.")
         @Size(groups = {SystemRoleManagementValidationGroup.Create.class},
                 min = 3, max = 100,
